Abstracted functions into Game Class
